refactor : 불필요한 메서드 삭제
