+++ src/main/resources/templates/chessGame.html
@@ -194,6 +194,8 @@
     document.getElementById("{{turn.key}}-player-turn").innerText = '현재 턴'
 
 
+
+
 </script>
 
 <script>
