+++ src/main/resources/templates/create.html
@@ -6,8 +6,8 @@
 </head>
 
 <body>
-<form action="/loadChessGame" method="post" id="load-form">
-    <input id="idInput" type="hidden" name="id" value="{{id}}">
+<form action="/loadChessGame" id="load-form" method="post">
+    <input id="idInput" name="id" type="hidden" value="{{id}}">
 </form>
 </body>
